Making a savespec file
